+++ pkg/rtc/mediatrackreceiver.go
@@ -17,6 +17,7 @@ import (
 	"github.com/livekit/livekit-server/pkg/sfu"
 	"github.com/livekit/livekit-server/pkg/sfu/buffer"
 	"github.com/livekit/livekit-server/pkg/telemetry"
+	"github.com/livekit/livekit-server/pkg/utils"
 )
 
 const (
@@ -91,7 +92,7 @@ func (t *MediaTrackReceiver) Restart() {
 	t.lock.Unlock()
 
 	if receiver != nil {
-		receiver.SetMaxExpectedSpatialLayer(SpatialLayerForQuality(livekit.VideoQuality_HIGH))
+		receiver.SetMaxExpectedSpatialLayer(utils.SpatialLayerForQuality(livekit.VideoQuality_HIGH))
 		t.MediaTrackSubscriptions.Restart()
 	}
 }
@@ -391,27 +392,12 @@ func (t *MediaTrackReceiver) OnSubscribedMaxQualityChange(f func(trackID livekit
 		}
 		receiver := t.Receiver()
 		if receiver != nil {
-			receiver.SetMaxExpectedSpatialLayer(SpatialLayerForQuality(maxSubscribedQuality))
+			receiver.SetMaxExpectedSpatialLayer(utils.SpatialLayerForQuality(maxSubscribedQuality))
 		}
 	})
 }
 
 // ---------------------------
-
-func SpatialLayerForQuality(quality livekit.VideoQuality) int32 {
-	switch quality {
-	case livekit.VideoQuality_LOW:
-		return 0
-	case livekit.VideoQuality_MEDIUM:
-		return 1
-	case livekit.VideoQuality_HIGH:
-		return 2
-	case livekit.VideoQuality_OFF:
-		return -1
-	default:
-		return -1
-	}
-}
 
 func QualityForSpatialLayer(layer int32) livekit.VideoQuality {
 	switch layer {
