+++ pkg/rtc/subscribedtrack.go
@@ -11,6 +11,7 @@ import (
 
 	"github.com/livekit/livekit-server/pkg/rtc/types"
 	"github.com/livekit/livekit-server/pkg/sfu"
+	"github.com/livekit/livekit-server/pkg/utils"
 )
 
 const (
@@ -45,7 +46,7 @@ func NewSubscribedTrack(params SubscribedTrackParams) *SubscribedTrack {
 	}
 
 	if !s.params.AdaptiveStream {
-		s.params.DownTrack.SetMaxSpatialLayer(SpatialLayerForQuality(livekit.VideoQuality_HIGH))
+		s.params.DownTrack.SetMaxSpatialLayer(utils.SpatialLayerForQuality(livekit.VideoQuality_HIGH))
 	}
 	return s
 }
@@ -124,7 +125,7 @@ func (t *SubscribedTrack) UpdateVideoLayer() {
 	if settings.Width > 0 {
 		quality = t.MediaTrack().GetQualityForDimension(settings.Width, settings.Height)
 	}
-	t.DownTrack().SetMaxSpatialLayer(SpatialLayerForQuality(quality))
+	t.DownTrack().SetMaxSpatialLayer(utils.SpatialLayerForQuality(quality))
 }
 
 func (t *SubscribedTrack) updateDownTrackMute() {
